pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datalens-ai"
dynamic = ["version"]
description = "AI-powered data analytics agent — upload data, ask questions in plain English, get insights"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Mekala Ajay", email = "mekalaajayk@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database :: Front-Ends",
]
keywords = ["ai", "data-analytics", "natural-language", "sql", "duckdb", "visualization"]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "plotly>=5.0",
    "duckdb>=1.0",
    "rich>=13.0",
    "httpx>=0.25",
    "tenacity>=8.0",
    "scikit-learn>=1.3",
    "openpyxl>=3.1",
    "jinja2>=3.1",
]

[project.optional-dependencies]
streamlit = ["streamlit>=1.30"]
gemini = ["google-generativeai>=0.4"]
pdf = ["xhtml2pdf>=0.2.11"]
all = [
    "datalens-ai[streamlit,gemini,pdf]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocs-minify-plugin>=0.8",
]

[project.scripts]
datalens = "datalens_ai.cli:main"

[project.urls]
Homepage = "https://github.com/mekala27-45/datalens-ai"
Documentation = "https://mekala27-45.github.io/datalens-ai/"
Repository = "https://github.com/mekala27-45/datalens-ai"
Issues = "https://github.com/mekala27-45/datalens-ai/issues"

[tool.hatch.version]
path = "src/datalens_ai/__init__.py"

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
